backend/requirements.txt
arrow==0.14.2
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dateparser==0.7.1
dj-database-url==0.5.0
Django==2.0
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
moment==0.8.2
openapi-codec==1.3.2
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
regex==2019.6.8
requests==2.22.0
simplejson==3.16.0
six==1.12.0
times==0.7
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
